refactor: convert map_if.py to a package; move x12_node to _base
Step 3 of the map_if refactor plan. The 1633-line module becomes a
package directory so subsequent PRs can move classes into siblings
without churn-by-Python-import-system. This PR is mechanical only —
identical runtime behavior, identical public surface.

Layout change:
  pyx12/map_if.py
    -> pyx12/map_if/__init__.py        (the rest, with adjusted
                                        relative imports `..X`)
       pyx12/map_if/_base.py           (x12_node + _required_attr
                                        + MAXINT)

x12_node, _required_attr (used by 4 of 5 node __init__ methods to
parse XML attrs), and the MAXINT constant move to _base. They are
the smallest, most self-contained piece — testing the package shim
mechanic before larger moves in PR 4 and PR 5.

The package __init__.py re-exports everything that was previously
top-level on `pyx12.map_if`, so all external callers and tests
keep working unchanged: `from pyx12.map_if import load_map_file`,
`pyx12.map_if.x12_node`, etc. Smoke check verified locally.

pyx12.pyproj (Visual Studio project file) updated from one
single-file `<Compile Include>` to two entries pointing at the
new submodule paths.

Verification gates all pass:
- pytest: 536 passed
- mypy --strict: clean
- ruff format + ruff check --select I: clean
- import smoke: all of load_map_file, map_if, x12_node, loop_if,
  segment_if, element_if, composite_if, _required_attr, MAXINT
  importable from pyx12.map_if

Net file count: +1 (one file split into two).
Net behavior change: zero.
