+++ pyx12.pyproj
@@ -69,7 +69,8 @@
     <Compile Include="pyx12\error_html.py" />
     <Compile Include="pyx12\error_item.py" />
     <Compile Include="pyx12\error_visitor.py" />
-    <Compile Include="pyx12\map_if.py" />
+    <Compile Include="pyx12\map_if\__init__.py" />
+    <Compile Include="pyx12\map_if\_base.py" />
     <Compile Include="pyx12\map_index.py" />
     <Compile Include="pyx12\map_override.py" />
     <Compile Include="pyx12\map_walker.py" />
